+++ .vscode/settings.json
@@ -14,6 +14,6 @@
     // Turn off tsc task auto detection since we have the necessary tasks as npm scripts
     "typescript.tsc.autoDetect": "off",
     "editor.formatOnSave": true,
-    "cSpell.words": ["codicon", "guids"]
+    "cSpell.words": ["codicon", "codicons", "guids", "tsbuildinfo"]
 
 }
