+++ src/extension/package.json
@@ -99,5 +99,8 @@
                 }
             ]
         }
+    },
+    "dependencies": {
+        "@t200logs/common": "^workspace:*"
     }
 }
