+++ src/ui/package.json
@@ -19,6 +19,7 @@
         "build": "webpack --config webpack.config.js --progress --mode production"
     },
     "dependencies": {
+        "@t200logs/common": "^workspace:*",
         "@vscode/codicons": "^0.0.35",
         "@vscode/webview-ui-toolkit": "^1.4.0",
         "react": "^18.2.0",
